feat: Improved database structure and currency handling

+++ Economy.Contract/src/IEconomyAPIv1.cs
@@ -6,29 +6,29 @@ public interface IEconomyAPIv1
 {
     void EnsureWalletKind(string kindName);
 
-    int GetPlayerBalance(IPlayer player, string walletKind);
-    int GetPlayerBalance(int playerid, string walletKind);
-    int GetPlayerBalance(ulong steamid, string walletKind);
+    decimal GetPlayerBalance(IPlayer player, string walletKind);
+    decimal GetPlayerBalance(int playerid, string walletKind);
+    decimal GetPlayerBalance(ulong steamid, string walletKind);
 
-    void SetPlayerBalance(IPlayer player, string walletKind, int amount);
-    void SetPlayerBalance(int playerid, string walletKind, int amount);
-    void SetPlayerBalance(ulong steamid, string walletKind, int amount);
+    void SetPlayerBalance(IPlayer player, string walletKind, decimal amount);
+    void SetPlayerBalance(int playerid, string walletKind, decimal amount);
+    void SetPlayerBalance(ulong steamid, string walletKind, decimal amount);
 
-    void AddPlayerBalance(IPlayer player, string walletKind, int amount);
-    void AddPlayerBalance(int playerid, string walletKind, int amount);
-    void AddPlayerBalance(ulong steamid, string walletKind, int amount);
+    void AddPlayerBalance(IPlayer player, string walletKind, decimal amount);
+    void AddPlayerBalance(int playerid, string walletKind, decimal amount);
+    void AddPlayerBalance(ulong steamid, string walletKind, decimal amount);
 
-    void SubtractPlayerBalance(IPlayer player, string walletKind, int amount);
-    void SubtractPlayerBalance(int playerid, string walletKind, int amount);
-    void SubtractPlayerBalance(ulong steamid, string walletKind, int amount);
+    void SubtractPlayerBalance(IPlayer player, string walletKind, decimal amount);
+    void SubtractPlayerBalance(int playerid, string walletKind, decimal amount);
+    void SubtractPlayerBalance(ulong steamid, string walletKind, decimal amount);
 
-    bool HasSufficientFunds(IPlayer player, string walletKind, int amount);
-    bool HasSufficientFunds(int playerid, string walletKind, int amount);
-    bool HasSufficientFunds(ulong steamid, string walletKind, int amount);
+    bool HasSufficientFunds(IPlayer player, string walletKind, decimal amount);
+    bool HasSufficientFunds(int playerid, string walletKind, decimal amount);
+    bool HasSufficientFunds(ulong steamid, string walletKind, decimal amount);
 
-    void TransferFunds(IPlayer fromPlayer, IPlayer toPlayer, string walletKind, int amount);
-    void TransferFunds(int fromPlayerid, int toPlayerid, string walletKind, int amount);
-    void TransferFunds(ulong fromSteamid, ulong toSteamid, string walletKind, int amount);
+    void TransferFunds(IPlayer fromPlayer, IPlayer toPlayer, string walletKind, decimal amount);
+    void TransferFunds(int fromPlayerid, int toPlayerid, string walletKind, decimal amount);
+    void TransferFunds(ulong fromSteamid, ulong toSteamid, string walletKind, decimal amount);
 
     void SaveData(IPlayer player);
     void SaveData(int playerid);
@@ -38,8 +38,8 @@ public interface IEconomyAPIv1
 
     bool WalletKindExists(string kindName);
 
-    event Action<ulong, string, long, long>? OnPlayerBalanceChanged;
-    event Action<ulong, ulong, string, long>? OnPlayerFundsTransferred;
+    event Action<ulong, string, decimal, decimal>? OnPlayerBalanceChanged;
+    event Action<ulong, ulong, string, decimal>? OnPlayerFundsTransferred;
     event Action<IPlayer>? OnPlayerLoad;
     event Action<IPlayer>? OnPlayerSave;
 }
